docker/requirements.txt
# requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
python-dotenv==0.19.2
transformers
git+https://github.com/aiola-lab/whisper-medusa.git
